move the Android test configuration into it's own source file
